Move CardState to types; add variant pure functions
